Add files via upload
In this project I aim to find out which of the three most popular startup books has the most in common with HBO's popular Silicon Valley TV show.  I do this by analyzing the most frequent and representative words and phrases of the books, define topics and compare them to the text of the show.